-Added assets folder -Added assets to yaml -Refactored splash to onboarding
