+++ lib/core/routes/app_router.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 
 import '../../features/home/presentation/home_view.dart';
-import '../../features/splash/presentation/views/splash_view.dart';
+import '../../features/onboarding/presentation/views/onboarding_view.dart';
 
 class AppRouter {
-  static const String splash = '/splash';
+  static const String onboarding = '/onboarding';
   static const String home = '/home';
 
   static Route<dynamic>? onGenerateRoute(RouteSettings settings) {
     switch (settings.name) {
-      case splash:
+      case onboarding:
         return MaterialPageRoute(builder: (context) => const SplashView());
       case home:
         return MaterialPageRoute(builder: (context) => const HomeView());
