+++ lib/talent_hub_app.dart
@@ -15,6 +15,6 @@ class TalentHubApp extends StatelessWidget {
             title: 'Talent Hub',
             debugShowCheckedModeBanner: false,
             onGenerateRoute: AppRouter.onGenerateRoute,
-            initialRoute: AppRouter.splash));
+            initialRoute: AppRouter.onboarding));
   }
 }
